Use config and wrapper to provide proxy configurations (#52)
* Add option for specifying a mask

Added the optional argument "mask" to the query_netcdf() function and subsequently to the parse_time_series_params() function. This allows a user to download masked netcdf files. First tests show good results. If mask="Land" then the oceans are masked by -666.

* Add proxies to config.

Users can specify the "PROXIES" in the Config. The wrapper will make sure that the proxies are forwarded to the requests get() and post() requests.

* Adding example for using the proxy

* Align unrelated changes with meteomatics master

* Adding comment to handle_proxy()

Co-authored-by: bbuman <bbuman@meteomatics.com>